removed shooter angle subsystem and moved it into shooter subsystem
