+++ src/main/java/org/usfirst/frc/team1360/robot/commands/ShooterCommand.java
@@ -13,7 +13,7 @@ public class ShooterCommand extends Command {
     public ShooterCommand() {
         // Use requires() here to declare subsystem dependencies
         // eg. requires(chassis);
-        requires(Subsystems.SHOOTER_ANGLE_SUBSYSTEM);
+        requires(Subsystems.SHOOTER_SUBSYSTEM);
     }
 
     // Called just before this Command runs the first time
@@ -22,7 +22,7 @@ public class ShooterCommand extends Command {
 
     // Called repeatedly when this Command is scheduled to run
     protected void execute() {
-        Subsystems.SHOOTER_ANGLE_SUBSYSTEM.changeShintakePosition(OI.getOperatorShooterSetpointUp() ,OI.getOperatorShooterSetpointDown());
+        Subsystems.SHOOTER_SUBSYSTEM.changeShintakePosition(OI.getOperatorShooterSetpointUp() ,OI.getOperatorShooterSetpointDown());
     }
 
     // Make this return true when this Command no longer needs to run execute()
