+++ src/main/java/org/usfirst/frc/team1360/robot/util/Subsystems.java
@@ -1,6 +1,5 @@
 package org.usfirst.frc.team1360.robot.util;
 
-import org.usfirst.frc.team1360.robot.subsystems.ShooterAngleSubsystem;
 import org.usfirst.frc.team1360.robot.subsystems.ShooterSubsystem;
 
 import org.usfirst.frc.team1360.robot.subsystems.DriveSubsystem;
@@ -13,5 +12,4 @@ public class Subsystems
      */
     public static final DriveSubsystem DRIVE_SUBSYSTEM = new DriveSubsystem();
     public static final ShooterSubsystem SHOOTER_SUBSYSTEM = new ShooterSubsystem();
-    public static final ShooterAngleSubsystem SHOOTER_ANGLE_SUBSYSTEM = new ShooterAngleSubsystem();
 }
